safari-600.7-branch/LayoutTests/platform/ios-sim/http/tests/security/directive-parsing-expected.txt

CONSOLE MESSAGE: line 1: Refused to load script from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js' because of Content-Security-Policy.

CONSOLE MESSAGE: line 1: Refused to load script from 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js' because of Content-Security-Policy.

None of these scripts should execute even though there are parse errors in the policy.

   

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
FAIL
